Separate util file for test utils
